+++ src/panels/doc_view.cpp
@@ -1,0 +1,186 @@
+/////////////////////////////////////////////////////////////////////////////
+// Purpose:   Panel for displaying docs in wxWebView
+// Author:    Ralph Walden
+// Copyright: Copyright (c) 2023 KeyWorks Software (Ralph Walden)
+// License:   Apache License -- see ../../LICENSE
+/////////////////////////////////////////////////////////////////////////////
+
+#include <wx/webview.h>
+
+#if !wxUSE_WEBVIEW_WEBKIT && !wxUSE_WEBVIEW_WEBKIT2 && !wxUSE_WEBVIEW_IE && !wxUSE_WEBVIEW_EDGE
+    #error "A wxWebView backend is required!"
+#endif
+
+#include "doc_view.h"
+
+#include "base_generator.h"   // BaseGenerator -- Base widget generator class
+#include "cstm_event.h"       // CustomEvent -- Custom Event class
+#include "mainframe.h"        // MainFrame -- Main window frame
+#include "node.h"             // Node class
+#include "project_handler.h"  // ProjectHandler -- Project file handler
+
+DocViewPanel::DocViewPanel(wxWindow* parent, MainFrame* frame) : DocViewBase(parent)
+{
+    m_mainframe = frame;
+    // Don't allow events to propagate any higher than this window.
+    SetExtraStyle(wxWS_EX_BLOCK_EVENTS);
+
+    frame->AddCustomEventHandler(GetEventHandler());
+
+    Bind(EVT_NodeSelected, &DocViewPanel::OnNodeSelected, this);
+}
+
+void DocViewPanel::ActivatePage()
+{
+    if (!m_webview)
+    {
+        wxBusyCursor wait;
+
+        m_language = Project.get_PreferredLanguage();
+        m_toolBar->ToggleTool(ID_CPLUS, m_language == GEN_LANG_CPLUSPLUS);
+        m_toolBar->ToggleTool(ID_PYTHON, m_language == GEN_LANG_PYTHON);
+        m_toolBar->ToggleTool(ID_RUBY, m_language == GEN_LANG_RUBY);
+
+        m_webview = wxWebView::New(this, wxID_ANY, "about:blank");
+        m_parent_sizer->Add(m_webview, wxSizerFlags(1).Expand().Border(wxALL));
+        m_parent_sizer->Layout();
+    }
+
+    wxCommandEvent dummy;
+    switch (m_language)
+    {
+        case GEN_LANG_CPLUSPLUS:
+            OnCPlus(dummy);
+            break;
+        case GEN_LANG_PYTHON:
+            OnPython(dummy);
+            break;
+        case GEN_LANG_RUBY:
+            OnRuby(dummy);
+            break;
+    }
+}
+
+void DocViewPanel::OnNodeSelected(CustomEvent& WXUNUSED(event))
+{
+    if (!IsShown())
+        return;
+
+    ActivatePage();
+}
+
+void DocViewPanel::OnCPlus(wxCommandEvent& /* event */)
+{
+    wxBusyCursor wait;
+    m_language = GEN_LANG_CPLUSPLUS;
+    if (auto* cur_sel = m_mainframe->GetSelectedNode(); cur_sel)
+    {
+        if (auto* gen = cur_sel->GetGenerator(); gen)
+        {
+            if (auto file = gen->GetHelpURL(cur_sel); file.size())
+            {
+                wxString url("https://docs.wxwidgets.org/3.2.0/class");
+                url << file.make_wxString();
+
+                m_webview->LoadURL(url);
+                return;
+            }
+        }
+    }
+    m_webview->SetPage("<html><title>Select Node</title>"
+                       "<body>The selected node does not have any specific documentation for this language.</body></html>",
+                       wxEmptyString);
+}
+
+void DocViewPanel::OnPython(wxCommandEvent& /* event */)
+{
+    wxBusyCursor wait;
+    m_language = GEN_LANG_PYTHON;
+    if (auto* cur_sel = m_mainframe->GetSelectedNode(); cur_sel)
+    {
+        if (auto* gen = cur_sel->GetGenerator(); gen)
+        {
+            if (auto file = gen->GetPythonURL(cur_sel); file.size())
+            {
+                wxString url("https://docs.wxpython.org/");
+                url << file.make_wxString();
+
+                m_webview->LoadURL(url);
+                return;
+            }
+        }
+    }
+    m_webview->SetPage("<html><title>Select Node</title>"
+                       "<body>The selected node does not have any specific documentation for this language.</body></html>",
+                       wxEmptyString);
+}
+
+void DocViewPanel::OnRuby(wxCommandEvent& /* event */)
+{
+    wxBusyCursor wait;
+    m_language = GEN_LANG_RUBY;
+    if (auto* cur_sel = m_mainframe->GetSelectedNode(); cur_sel)
+    {
+        if (auto* gen = cur_sel->GetGenerator(); gen)
+        {
+            if (auto file = gen->GetRubyURL(cur_sel); file.size())
+            {
+                wxString url("https://mcorino.github.io/wxRuby3/");
+                url << file.make_wxString();
+
+                m_webview->LoadURL(url);
+                return;
+            }
+        }
+    }
+    m_webview->SetPage("<html><title>Select Node</title>"
+                       "<body>The selected node does not have any specific documentation for this language.</body></html>",
+                       wxEmptyString);
+}
+
+void DocViewPanel::OnHome(wxCommandEvent& WXUNUSED(event))
+{
+    ActivatePage();
+}
+
+void DocViewPanel::OnBack(wxCommandEvent& WXUNUSED(event))
+{
+    wxBusyCursor wait;
+
+    m_webview->GoBack();
+}
+
+void DocViewPanel::OnForward(wxCommandEvent& WXUNUSED(event))
+{
+    wxBusyCursor wait;
+    m_webview->GoForward();
+}
+
+void DocViewPanel::OnUpdateBack(wxUpdateUIEvent& event)
+{
+    if (!m_webview)
+    {
+        event.Enable(false);
+        return;
+    }
+
+    if (auto history = m_webview->GetBackwardHistory(); history.size() == 1)
+    {
+        event.Enable(history[0].get()->GetUrl() != "about:blank");
+        return;
+    }
+
+    event.Enable(m_webview->CanGoBack());
+}
+
+void DocViewPanel::OnUpdateForward(wxUpdateUIEvent& event)
+{
+    if (!m_webview)
+    {
+        event.Enable(false);
+
+        return;
+    }
+
+    event.Enable(m_webview->CanGoForward());
+}